backend/requirements.txt
fastapi
uvicorn
bcrypt
pydantic
passlib
pyjwt